refactor: formatter 구현부를 Double의 extension에서 BankManager의 메서드로 수정, 에러메시지 출력 시 localizedDescription 활용, BankManager의 timePerTask 프로퍼티 삭제
