+++ BankManagerConsoleApp/BankManagerConsoleApp/BankError.swift
@@ -2,16 +2,21 @@
 //  BankError.swift
 //  BankManagerConsoleApp
 
+import Foundation
+
 enum BankError: Error {
     case invalidInput
     case formattingError
     
-    var message: String {
+}
+
+extension BankError: LocalizedError {
+    var errorDescription: String? {
         switch self {
         case .invalidInput:
-            return "잘못된 입력"
+            return NSLocalizedString("잘못된 입력", comment: "invalid input")
         case .formattingError:
-            return "포맷팅 실패"
+            return NSLocalizedString("포맷팅 실패", comment: "formatting error")
         }
     }
 }
