+++ BankManagerConsoleApp/BankManagerConsoleApp/Constant.swift
@@ -6,4 +6,5 @@ enum Constant {
     static let openBank: String = "1"
     static let closeBank: String = "2"
     static let customerNumberRange: ClosedRange<Int> = 10...30
+    static let micro: Double = 0.000001
 }
